Refactor to simplify index file.
